requirements.txt
# OrbitMind Dependencies

# Core - Data Collection
lightstreamer-client-lib>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# ML (Phase 2)
# river>=0.21.0
# pandas>=2.0.0
# numpy>=1.24.0

# Dashboard (Phase 3)
# streamlit>=1.30.0
# plotly>=5.18.0

# Dev
# pytest>=7.0.0
# ipython>=8.0.0
